Removes type of public imports from types scope
If a file nicknamed 'im1' is imported, and if 'im1' also imports another file 'im2' as public, then the type corresponding to 'im2' is necessary when constructing the method in 'im1' that allows us to access the methods in 'im2'.
This change removes the type belonging to 'im2' from the types scope after creating this method since it was only meant to be temporary.

Co-Authored-By: Joe Brennan <jwsbrennan@gmail.com>