Testing ampersand and subtyping
Adds several new tests for making sure ampersand is correctly evaluated.
Adds a test for the infinite loop problem with trying to subtype self-referential methods in types that have been &'d together.
Moves tests that used to be in the 'Broken' folder into the main testing folder, as they are now fixed.
Deletes the rest of the tests for embedded types.

Co-Authored-By: vanthuan9 <thuan.nguyen@pomona.edu>