ObjectTypes not defined by 'ast.opNode' don't store their AstNode
Creating an ObjectType no longer requires its corresponding AstNode. ObjectTypes
now have the methods: 'isOp', 'getOpNode', and 'setOpNode' to facilitate in
subtyping of types defined by 'ast.opNode'.
Removed the type MethodPair as it is now unnecessary for ‘fromGctLine’ to construct and return the methodTypeNode corresponding to the gct line.
Added new type TypeOp to more easily store the relevant information that was previously in the AstNode.

Co-Authored-By: vanthuan9 <thuan.nguyen@pomona.edu>